Rename logilica_weekly_report directory to logilica_cli
