+++ pyproject.toml
@@ -11,7 +11,7 @@ authors = [{ name = "The Developer Practices Team", email = "developer-productiv
 license = {"text" = "Apache-2.0"}
 
 [project.scripts]
-logilica-weekly-report = "logilica_weekly_report.__main__:cli"
+logilica-weekly-report = "logilica_cli.__main__:cli"
 
 [tool.autoflake]
 exclude = ["build"]
@@ -35,7 +35,7 @@ wrap-summaries = 79
 [tool.coverage.run]
 branch = true
 relative_files = true
-source = ["logilica_weekly_report", "tests"]
+source = ["logilica_cli", "tests"]
 
 [tool.coverage.report]
 # Regexes for lines to exclude from consideration
@@ -61,10 +61,10 @@ skip_empty = true
 skip_empty = true
 
 [tool.isort]
-profile = "black"                               # black-compatible (e.g., trailing comma)
-known_first_party = ["logilica_weekly_report"]  # use separate section for project sources
-force_sort_within_sections = true               # don't separate import vs from
-order_by_type = false                           # sort alphabetic regardless of case
+profile = "black"                     # black-compatible (e.g., trailing comma)
+known_first_party = ["logilica_cli"]  # use separate section for project sources
+force_sort_within_sections = true     # don't separate import vs from
+order_by_type = false                 # sort alphabetic regardless of case
 
 [tool.pytest.ini_options]
 filterwarnings = [
@@ -76,7 +76,7 @@ empty_parameter_set_mark = "xfail"
 testpaths = ["tests"]
 
 [tool.setuptools]
-packages = ["logilica_weekly_report"]
+packages = ["logilica_cli"]
 
 [tool.setuptools.dynamic]
 dependencies = { file = "requirements.txt" }
